ref(presign-api): consolidate validators into controller
- shift validators into admin-form controller
- update handler methods to include validators as request handler array
- update controller integration tests
- update old routes to use new handler
- update new routes to use new handleri
